rename main file to index in favorite page modules
